fix: address AStaroverov review — brand log handles, backoff, move tokens logic
- Brand LiveLogHandle/ReadyLogHandle with Branded<> (#2)
- Add 500ms minimum delay between polls in await_block_done (#3)
- Remove duplicate AuthorMarker, import from pl-middle-layer (#4)
- Move estimateTokens/summarizeOutputs from types.ts to tokens.ts (#6)
- Tests already exist and pass from tokens.ts (#7)
